+++ src/configs/index.ts
@@ -1,6 +1,0 @@
-export * from './jest-dom.js'
-export * from './jest.js'
-export * from './next.js'
-export * from './node.js'
-export * from './react.js'
-export * from './vitest.js'