+++ src/configs/jest-dom.spec.ts
@@ -1,7 +1,5 @@
 import { sharedPlugins } from '../utils/constants.js'
-import { jestDomConfig } from './jest-dom.js'
-
-const sut = await jestDomConfig()
+import { jestDomConfig as sut } from './jest-dom.js'
 
 describe('jest-dom', () => {
   it('should include shared plugins', () => {