+++ src/index.ts
@@ -1,5 +1,22 @@
-export * from './configs/index.js'
-export * from './factory.js'
+import { version as packageVersion } from '../package.json'
+import { jestDomConfig } from './configs/jest-dom.js'
+import { jestConfig } from './configs/jest.js'
+import { nextConfig } from './configs/next.js'
+import { nodeConfig } from './configs/node.js'
+import { reactConfig } from './configs/react.js'
+import { vitestConfig } from './configs/vitest.js'
 
-// eslint-disable-next-line import/no-default-export
-export { config as default } from './factory.js'
+export = {
+  meta: {
+    name: 'mist3rbru',
+    version: packageVersion,
+  },
+  configs: {
+    jest: jestConfig,
+    react: reactConfig,
+    next: nextConfig,
+    node: nodeConfig,
+    vitest: vitestConfig,
+    'jest-dom': jestDomConfig,
+  },
+}