+++ src/plugins/typescript.ts
@@ -406,6 +406,15 @@ export const typescriptPlugin = definePlugin({
 
     // https://typescript-eslint.io/rules/consistent-return
     '@typescript-eslint/consistent-return': 'off',
+
+    // https://typescript-eslint.io/rules/use-unknown-in-catch-callback-variable
+    '@typescript-eslint/use-unknown-in-catch-callback-variable': 'off',
+
+    // https://typescript-eslint.io/rules/restrict-plus-operands
+    '@typescript-eslint/restrict-plus-operands': [
+      'error',
+      { allowNumberAndString: true },
+    ],
   },
   testRules: {
     // https://typescript-eslint.io/rules/no-explicit-any