+++ src/plugins/vitest.spec.ts
@@ -1,7 +1,7 @@
 import type { EslintRuleMeta } from '../types.js'
 import { vitestPlugin as sut } from './vitest.js'
 import { readFile } from 'node:fs/promises'
-import { resolve } from 'node:path'
+import path from 'node:path'
 import _vitestPlugin from 'eslint-plugin-vitest'
 
 describe('vitest', () => {
@@ -18,7 +18,7 @@ describe('vitest', () => {
   })
 
   it("should include rule's reference link", async () => {
-    const file = await readFile(resolve('src/plugins/vitest.ts'))
+    const file = await readFile(path.resolve('src/plugins/vitest.ts'))
     const fileContent = file.toString()
 
     const expectedReferencedRules = fileContent